Issue-37 - Added a simple type='colorbox' display field, and type='color' input field
1) Added support for <input type='color'>

Added support for HTML <input type='color'> fields, which yield an #RRGGBB hex string, and fall back to simple text input on unsupported browsers.

2) Added new display type type='colorbox'

This renders an #RRGGBB color string in a rectangle, overlaid with the color string (XOR-ed with the color, so the text is always visible).

3) Updated README

Documented the new color input and display fields.
Also documented the <input type='email'> field, which was missing.